Former-commit-id: 1a30688ffa36810c8b39daaf21db3a45173d3ecf

+++ Sources/Internal/Entity/BaseFamily.h
@@ -100,7 +100,7 @@ BaseFamily<Component>::BaseFamily(const Vector<Component*>& components)
         uint32 type = components[i]->GetType();
         componentIndices[type] = i;
         componentCount[type]++;
-        componentsFlags |= (uint64)1 << type;
+        componentsFlags |= 1ULL << type;
     }
 }
